\name{gghexbin-package}
\alias{gghexbin-package}
\alias{gghexbin}
\docType{package}
\title{\packageTitle{gghexbin}}
\description{\packageDescription{gghexbin}}
\section{Package Content}{\packageIndices{gghexbin}}
\author{\packageAuthor{gghexbin}}
\section{Maintainer}{\packageMaintainer{gghexbin}}
\keyword{package}
