% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{hex}
\alias{hex}
\title{Something about what hex.R does}
\usage{
hex(
  x,
  y,
  xmin = min(x),
  xmax = max(x),
  ymin = min(y),
  ymax = max(y),
  pct_of_plot_used = 20,
  binning_relation = 0.5,
  lattice = TRUE,
  plotted = TRUE
)
}
\arguments{
\item{x}{add details}

\item{y}{add details}
}
\value{
Something about what is returned using for example \code{x} and \code{y}
}
\description{
Something about what hex.R does
}
\examples{
hex(x = seq(1:10), y = seq(1:10))
}
