% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{gghex}
\alias{gghex}
\title{Something about what hex.R does}
\usage{
gghex(x, y, pct_of_plot_used = 20, binning_relation = 0.5)
}
\arguments{
\item{x}{independent variable}

\item{y}{dependent variable}

\item{pct_of_plot_used}{maximum value of y, default value is max(y)}

\item{binning_relation}{a value between 0 and 1 that determines the relationship between count and hexagon area proportion. 
A value of 0.5 means that area and count have a 1:1 ratio, A value of 0 means that all hexagons are the same size, regardless of count}
}
\value{
Returns lattice information, \code{lt}, that includes hexagons' circumcenter coordinates, radii, and counts; 
Also returns ggproto object, \code{hexplot}, which contains all of the geom_poly() generated hexagons for plotting.
}
\description{
Something about what hex.R does
}
\examples{
library(ggplot2)
library(data.table)

data(iris)
x=iris$Sepal.Length
y=iris$Petal.Length

lin_reg <- lm(y~x, data.table(x,y))
ggplot() + gghex(x, y, pct_of_plot_used = 20, binning_relation = 0.5) + ggtitle("iris Data") + xlab("Sepal Length") + ylab("Petal Length") + geom_abline(slope = lin_reg$coefficients[2], intercept = lin_reg$coefficients[1]) + coord_fixed()  + guides(fill=guide_legend(title="Counts")) + scale_fill_gradient(low="green", high="red")

}
